address review comments and refactor
